improved the overall code organization
