+++ app/desmos/cmd/chainlink/builder/provider/provider.go
@@ -1,0 +1,17 @@
+package provider
+
+import (
+	"github.com/desmos-labs/desmos/v2/app/desmos/cmd/chainlink/builder"
+	multibuilder "github.com/desmos-labs/desmos/v2/app/desmos/cmd/chainlink/builder/multi"
+	singlebuilder "github.com/desmos-labs/desmos/v2/app/desmos/cmd/chainlink/builder/single"
+	multigetter "github.com/desmos-labs/desmos/v2/app/desmos/cmd/chainlink/getter/multi"
+	singlegetter "github.com/desmos-labs/desmos/v2/app/desmos/cmd/chainlink/getter/single"
+)
+
+// DefaultChainLinkJSONBuilderProvider returns the default ChainLinkJSONBuilder provider implementation
+func DefaultChainLinkJSONBuilderProvider(isSingleAccount bool) builder.ChainLinkJSONBuilder {
+	if isSingleAccount {
+		return singlebuilder.NewAccountChainLinkJSONBuilder(singlegetter.NewChainLinkJSONInfoGetter())
+	}
+	return multibuilder.NewAccountChainLinkJSONBuilder(multigetter.NewChainLinkJSONInfoGetter())
+}
